move hook to component
