--- packages/mobile/src/hooks/useLoadingAnimation.ts
@@ -1,17 +1,0 @@
-import { useRef, useLayoutEffect } from 'react'
-
-import { LayoutAnimation } from 'react-native'
-
-const useLoadingAnimation = (isDepLoaded: () => boolean, dependency: any) => {
-  // Prevents multiple re-renders if the dependency changes.
-  const isLoaded = useRef(false)
-
-  useLayoutEffect(() => {
-    if (isDepLoaded() && !isLoaded.current) {
-      isLoaded.current = true
-      LayoutAnimation.configureNext(LayoutAnimation.Presets.easeInEaseOut)
-    }
-  }, [dependency, isDepLoaded])
-}
-
-export default useLoadingAnimation
